+++ lib/widgets/select_multi_currency_sheet.dart
@@ -71,7 +71,7 @@ class _SelectMultiCurrencySheetState extends State<SelectMultiCurrencySheet> {
                 final CurrencyData transformedCurrencyData =
                     iso4217CurrenciesGrouped[currency.code]!;
 
-                return CheckboxListTile /*.adaptive*/ (
+                return CheckboxListTile(
                   value: selectedUuids.contains(transformedCurrencyData.code),
                   title: Text(transformedCurrencyData.name),
                   subtitle: Text(