+++ lib/widgets/transaction_filter_head/transaction_search_sheet.dart
@@ -55,7 +55,6 @@ class _TransactionSearchSheetState extends State<TransactionSearchSheet> {
           ),
         ],
       ),
-      scrollableContentMaxHeight: MediaQuery.of(context).size.height * 0.5,
       child: SingleChildScrollView(
         child: Column(
           mainAxisSize: MainAxisSize.min,
@@ -92,7 +91,7 @@ class _TransactionSearchSheetState extends State<TransactionSearchSheet> {
               ),
             ),
             const SizedBox(height: 16.0),
-            CheckboxListTile /*.adaptive*/ (
+            CheckboxListTile(
               title: Text(
                 "transactions.query.filter.keyword.includeDescription".t(
                   context,