+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Next
 
+### New features
+
+* Now you can save backups to iCloud. Enable it in `Preferences > Sync`
+
+### Changes & Improvements
+
+* Backup files naming convention has changed to comply with iCloud Drive file name requirements
 * Added "Recommend Flow" button
 
 ## 0.13.1